requirements.txt
torch >= 2.1.0
torchaudio >= 2.0.1
torchvision >= 0.16.0
python-dotenv >= 1.0.0
click >= 8.1.0
pydantic >= 2.5.3
hydra-core >= 1.3.0
lightning >= 2.1.0
pandas >= 2.1.4
wandb >= 0.12.10
matplotlib >= 3.8.2
